made the chat fragments into activities
