--- app/src/main/java/com/example/android/kfupmsocialspace/ChatFilesFragment.java
@@ -1,18 +1,0 @@
-package com.example.android.kfupmsocialspace;
-
-import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-public class ChatFilesFragment extends Fragment {
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        //inflates sets or shows the fragment selected in the first param
-        return inflater.inflate(R.layout.activity_chat_files_fragment, container, false);
-    }
-}
