+++ app/src/main/java/com/example/android/kfupmsocialspace/ChatInfoActivity.java
@@ -1,0 +1,13 @@
+package com.example.android.kfupmsocialspace;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+
+public class ChatInfoActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_chat_info);
+    }
+}
