+++ app/src/main/java/com/example/android/kfupmsocialspace/UtilitiesActivity.java
@@ -10,6 +10,8 @@ import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
@@ -55,7 +57,6 @@ public class UtilitiesActivity extends AppCompatActivity {
 
 
     //This part adds the three dots on the top right
-    /*
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
@@ -78,7 +79,6 @@ public class UtilitiesActivity extends AppCompatActivity {
         return super.onOptionsItemSelected(item);
     }
 
-    */
 
     @Override
     public void onBackPressed() {
