--- app/src/main/res/layout/activity_chat_chat_fragment.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ChatActivityChatFragment">
-
-    <FrameLayout
-        android:id="@+id/chat_messages_frame"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_marginBottom="?android:attr/actionBarSize">
-
-        <android.support.v7.widget.RecyclerView
-            android:id="@+id/messages_list"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-        </android.support.v7.widget.RecyclerView>
-    </FrameLayout>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:orientation="horizontal"
-        android:weightSum="1">
-
-        <ImageButton
-            android:id="@+id/attach_file"
-            android:layout_width="0dp"
-            android:layout_height="match_parent"
-            android:layout_margin="4dp"
-            android:layout_weight="0.1"
-            android:background="?android:attr/selectableItemBackground"
-            android:src="@drawable/ic_attach_file" />
-
-        <EditText
-            android:id="@+id/message_string"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="0.8"
-            android:focusedByDefault="false"
-            android:hint="Type a message"
-            android:textSize="20sp" />
-
-        <ImageButton
-            android:id="@+id/send_message"
-            android:layout_width="0dp"
-            android:layout_height="match_parent"
-            android:layout_margin="4dp"
-            android:layout_weight="0.1"
-            android:background="?android:attr/selectableItemBackground"
-            android:src="@drawable/ic_send" />
-    </LinearLayout>
-</RelativeLayout>
