+++ app/src/main/res/layout/activity_chat_files.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@android:color/holo_purple"
-    tools:context=".ChatFilesFragment">
+    tools:context=".ChatFilesActivity">
 
     <TextView
         android:layout_width="wrap_content"
