+++ app/src/main/res/layout/activity_chat_group.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@android:color/holo_blue_bright"
-    tools:context=".ChatGroupFragment">
+    tools:context=".ChatGroupActivity">
 
     <TextView
         android:layout_width="wrap_content"
