+++ app/src/main/res/menu/menu_utilities.xml
@@ -4,7 +4,7 @@
     tools:context="com.example.android.kfupmsocialspace.UtilitiesActivity">
     <item
         android:id="@+id/action_settings"
-        android:orderInCategory="100"
+
         android:title="@string/action_settings"
         app:showAsAction="never" />
 </menu>
